♻️ refactor: 파일명 수정 및 interface 분리
