+++ BE/src/stock/topfive/stock.topfive.service.ts
@@ -4,50 +4,14 @@ import { ConfigService } from '@nestjs/config';
 import { StockRankigRequestDto } from './dto/stock-ranking-request.dto';
 import { StockRankingResponseDto } from './dto/stock-ranking-response.dto';
 import { StockRankingDataDto } from './dto/stock-ranking-data.dto';
-
-export enum MarketType {
-  ALL = 'ALL',
-  KOSPI = 'KOSPI',
-  KOSDAQ = 'KOSDAQ',
-  KOSPI200 = 'KOSPI200',
-}
-
-interface StockApiOutputData {
-  stck_shrn_iscd: string;
-  data_rank: string;
-  hts_kor_isnm: string;
-  stck_prpr: string;
-  prdy_vrss: string;
-  prdy_vrss_sign: string;
-  prdy_ctrt: string;
-  acml_vol: string;
-  stck_hgpr: string;
-  hgpr_hour: string;
-  acml_hgpr_data: string;
-  stck_lwpr: string;
-  lwpr_hour: string;
-  acml_lwpr_date: string;
-  lwpr_vrss_prpr_rate: string;
-  dsgt_date_clpr_vrss_prpr_rate: string;
-  cnnt_ascn_dynu: string;
-  hgpr_vrss_prpr_rate: string;
-  cnnt_down_dynu: string;
-  oprc_vrss_prpr_sign: string;
-  oprc_vrss_prpr: string;
-  oprc_vrss_prpr_rate: string;
-  prd_rsfl: string;
-  prd_rsfl_rate: string;
-}
-
-interface StockApiResponse {
-  output: StockApiOutputData[];
-  rt_cd: string;
-  msg_cd: string;
-  msg1: string;
-}
+import {
+  StockApiOutputData,
+  StockApiResponse,
+} from './interface/stock.topfive.interface';
+import { MarketType } from '../enum/MarketType';
 
 @Injectable()
-export class TopFiveService {
+export class StockTopfiveService {
   private accessToken: string;
   private tokenExpireTime: Date;
   private readonly koreaInvestmentConfig: {
